TODO : Find a way to predict best wine
Since we're not working with grades, i might have an idea but i'll have to take a look a that later. It involves messing with the model so that in returns a probability associated with each prediction. If we take the prediction in the class "good" with the highest probability, we could consider that this is the best wine. Otherwise, I don't really know what to do using this method of classification